+++ src/Network/Network/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 --->
 ## Upcoming Release
+* Add support for Virtual Network Subnet Initialization
+    - New cmdlets
+        - Initialize-AzVirtualNetworkSubnetPolicy
 * Add support for Virtual Network Gateway Resource
     - New cmdlets
         - Get-AzVirtualNetworkGatewayVpnClientConnectionHealth